fix: Copilot追加レビュー指摘事項の修正
- CategorySummary.Totalをint64に変更してオーバーフローを根本的に解決しました
  - int型へのキャストによるオーバーフローを防ぎ、大きな金額でも安全に計算できます
- categorizeReceiptItemsのエラーハンドリングを改善しました
  - AI APIエラーやパースエラー時にデフォルトカテゴリー（その他）を設定します
  - カテゴリー判定の失敗がレシート保存を妨げないようになりました
- WebHandler.templatesフィールドを削除しました
  - リクエストごとにテンプレートをパースする方式に統一し、未使用フィールドを削除しました
- エラーハンドリング変更に対応したテストケースを追加・更新しました

+++ internal/modules/household/usecase/household_usecase.go
@@ -10,7 +10,7 @@ import (
 type CategorySummary struct {
 	Category string
 	Count    int
-	Total    int
+	Total    int64 // オーバーフロー対策のためint64を使用
 }
 
 // HouseholdUseCase 家計簿集計のユースケース
@@ -59,8 +59,8 @@ func (uc *HouseholdUseCase) GetCategorySummary(ctx context.Context) ([]CategoryS
 				}
 			}
 			summaryMap[category].Count++
-			// オーバーフロー対策：int64にキャストしてから計算
-			summaryMap[category].Total += int(int64(item.Price) * int64(item.Quantity))
+			// int64で安全に計算
+			summaryMap[category].Total += int64(item.Price) * int64(item.Quantity)
 		}
 	}
 
@@ -77,7 +77,7 @@ func (uc *HouseholdUseCase) GetCategorySummary(ctx context.Context) ([]CategoryS
 			}
 		}
 		summaryMap[expense.Category].Count++
-		summaryMap[expense.Category].Total += expense.Amount
+		summaryMap[expense.Category].Total += int64(expense.Amount)
 	}
 
 	// マップをスライスに変換
